Move Personal Candidate Functions
